requirements.txt
streamlit
pandas
scikit-learn
joblib
matplotlib
seaborn
transformers
sentence-transformers
faiss-cpu
torch 
huggingface-hub